Move DNB serializers to `dnb_api` package.
